infra/requirements.txt
attrs==22.2.0
aws-cdk-lib==2.45.0
boto3==1.26.85
botocore==1.29.85
black==23.3.0
